+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - periodically update drive selector
 - add a user's home to the drive selector on Windows
 - add Open in -> Turtle Commander action for Project files
+- calculate the directory size on its selection
 
 ## [0.2.1] - 2026-03-26